docker_compose/empty_project/movies_admin/requirements.txt
django==3.2
django-split-settings==1.2.0
flake8==4.0
python-dotenv==0.21.0
psycopg2-binary==2.9
uwsgi==2.0.20

